# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Appoint API",
    "description": "Appointment Management System API",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.appoint.local/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/owner/staff/{businessId}": {
      "post": {
        "tags": ["Staff"],
        "summary": "Create staff member",
        "operationId": "createStaff",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId"],
                "properties": {
                  "userId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Staff created"
          }
        }
      },
      "get": {
        "tags": ["Staff"],
        "summary": "Get business staff",
        "operationId": "getStaff",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of staff"
          }
        }
      }
    },
    "/owner/staff/{businessId}/{staffId}": {
      "patch": {
        "tags": ["Staff"],
        "summary": "Update staff (activate/deactivate)",
        "operationId": "updateStaff",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "staffId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Staff updated"
          }
        }
      }
    },
    "/owner/services/{businessId}": {
      "post": {
        "tags": ["Services"],
        "summary": "Create service",
        "operationId": "createService",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "duration", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "duration": { "type": "number" },
                  "price": { "type": "number" },
                  "staffIds": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created"
          }
        }
      },
      "get": {
        "tags": ["Services"],
        "summary": "Get services",
        "operationId": "getServices",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services"
          }
        }
      }
    },
    "/owner/services/{businessId}/{serviceId}": {
      "patch": {
        "tags": ["Services"],
        "summary": "Update service",
        "operationId": "updateService",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" },
                  "staffIds": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated"
          }
        }
      }
    },
...
    "/public/slots": {
      "get": {
        "tags": ["Slots"],
        "summary": "Get available slots",
        "operationId": "getSlots",
        "parameters": [
          {
            "name": "businessSlug",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "serviceId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "staffId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Available slots"
          }
        }
      }
    },
...
  }
}